Cleanup query logging code.
Remove code duplication
