+++ tests/TestCase/Listener/ApiQueryLogListenerTest.php
@@ -5,13 +5,10 @@ namespace Crud\Test\TestCase\Listener;
 
 use Cake\Controller\Controller;
 use Cake\Core\Configure;
-use Cake\Database\Connection;
-use Cake\Database\Driver;
 use Cake\Event\Event;
 use Cake\Http\ServerRequest;
 use Crud\Action\BaseAction;
 use Crud\Listener\ApiQueryLogListener;
-use Crud\Log\QueryLogger;
 use Crud\TestSuite\TestCase;
 
 /**
@@ -99,11 +96,11 @@ class ApiQueryLogListenerTest extends TestCase
         $Instance = $this
             ->getMockBuilder(ApiQueryLogListener::class)
             ->disableOriginalConstructor()
-            ->onlyMethods(['_getQueryLogs'])
+            ->onlyMethods(['getQueryLogs'])
             ->getMock();
         $Instance
             ->expects($this->never())
-            ->method('_getQueryLogs');
+            ->method('getQueryLogs');
 
         $Instance->beforeRender(new Event('something'));
     }
@@ -142,7 +139,7 @@ class ApiQueryLogListenerTest extends TestCase
         $Instance = $this
             ->getMockBuilder(ApiQueryLogListener::class)
             ->disableOriginalConstructor()
-            ->onlyMethods(['_getQueryLogs', '_action', '_controller'])
+            ->onlyMethods(['getQueryLogs', '_action', '_controller'])
             ->getMock();
         $Instance
             ->expects($this->once())
@@ -154,128 +151,13 @@ class ApiQueryLogListenerTest extends TestCase
             ->willReturn($Controller);
         $Instance
             ->expects($this->once())
-            ->method('_getQueryLogs')
+            ->method('getQueryLogs')
             ->willReturn([]);
 
         $Instance->beforeRender(new Event('something'));
     }
 
-    /**
-     * Test setting up the query loggers
-     *
-     * @return void
-     */
-    public function testSetupLogging()
-    {
-        $driver = $this
-            ->getMockBuilder(Driver::class)
-            ->getMock();
-        $driver
-            ->expects($this->once())
-            ->method('setLogger')
-            ->with($this->isInstanceOf(QueryLogger::class));
-        $driver
-            ->expects($this->any())
-            ->method('enabled')
-            ->willReturn(true);
-
-        $DefaultSource = new Connection([
-            'name' => 'default',
-            'driver' => $driver,
-        ]);
-
-        $Instance = $this
-            ->getMockBuilder(ApiQueryLogListener::class)
-            ->disableOriginalConstructor()
-            ->onlyMethods(['_getSources', '_getSource'])
-            ->getMock();
-        $Instance
-            ->expects($this->once())
-            ->method('_getSources')
-            ->willReturn(['default']);
-        $Instance
-            ->expects($this->any())
-            ->method('_getSource')
-            ->with('default')
-            ->willReturn($DefaultSource);
-
-        $Instance->setupLogging(new Event('something'));
-    }
-
-    /**
-     * Test setting up only specific query loggers
-     *
-     * @return void
-     */
-    public function testSetupLoggingConfiguredSources()
-    {
-        $driver = $this->getMockBuilder(Driver::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-        $driver
-            ->expects($this->any())
-            ->method('enabled')
-            ->willReturn(true);
-        $driver2 = $this->getMockBuilder(Driver::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-        $driver2
-            ->expects($this->any())
-            ->method('enabled')
-            ->willReturn(true);
-
-        $DefaultSource = new Connection([
-            'name' => 'default',
-            'driver' => $driver,
-        ]);
-        $TestSource = new Connection([
-            'name' => 'test',
-            'driver' => $driver2,
-        ]);
-
-        $Instance = $this
-            ->getMockBuilder(ApiQueryLogListener::class)
-            ->disableOriginalConstructor()
-            ->onlyMethods(['_getSources', '_getSource'])
-            ->getMock();
-        $Instance
-            ->expects($this->never())
-            ->method('_getSources');
-
-        $Instance
-            ->expects($this->exactly(2))
-            ->method('_getSource')
-            ->with(...self::withConsecutive(['default'], ['test']))
-            ->willReturnOnConsecutiveCalls($DefaultSource, $TestSource);
-
-        $Instance->setConfig('connections', ['default', 'test']);
-        $Instance->setupLogging(new Event('something'));
-    }
-
-    /**
-     * Test getting query logs using protected method
-     *
-     * @return void
-     */
-    public function testProtectedGetQueryLogs()
-    {
-        $listener = new ApiQueryLogListener(new Controller(new ServerRequest()));
-        $listener->setupLogging(new Event('something'));
-        $this->setReflectionClassInstance($listener);
-
-        $expected = [
-            'test' => [],
-        ];
-
-        $this->assertEquals($expected, $this->callProtectedMethod('_getQueryLogs', [], $listener));
-    }
-
-    /**
-     * Test getting query logs using public getter.
-     *
-     * @return void
-     */
-    public function testPublicGetQueryLogs()
+    public function testQueryLogs()
     {
         $listener = new ApiQueryLogListener(new Controller(new ServerRequest()));
         $listener->setupLogging(new Event('something'));
